backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
qdrant-client==1.7.0
cohere==4.37
tiktoken==0.5.2
pydantic==2.5.0
python-multipart==0.0.6
groq==0.4.1
google-generativeai==0.3.1
huggingface-hub==0.19.4
requests==2.31.0